Add Validation Class. And handle EmptyResultDataAccessException from UserService.

+++ src/main/java/com/spring/controller/UserController.java
@@ -3,6 +3,7 @@ package com.spring.controller;
 import com.spring.model.Message;
 import com.spring.model.User;
 import com.spring.service.UserService;
+import com.spring.util.Validation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,7 +11,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 import java.util.List;
 
 /**
@@ -29,27 +29,70 @@ public class UserController {
     }
 
     @RequestMapping(value = "/addUser", method = {RequestMethod.GET, RequestMethod.POST})
-    public ResponseEntity<Message> addUser(@Valid @ModelAttribute User user) {
+    public ResponseEntity<Message> addUser(@ModelAttribute User user) {
         logger.debug("Start addUser() for /addUser");
-        Message message = userService.addUser(user);
+        Validation validation = new Validation();
+        Message message = new Message();
+        if (validation.isValid(user)) {
+            message = userService.addUser(user);
+            return new ResponseEntity<Message>(message, HttpStatus.OK);
+        } else {
+            message.setMessage("empty/null input/s");
+            message.setStatus(false);
+            return new ResponseEntity<Message>(message, HttpStatus.BAD_REQUEST);
+        }
 
-        return new ResponseEntity<Message>(message, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/updateUser", method = RequestMethod.PUT)
-    public ResponseEntity<Message> updateUser(@Valid @ModelAttribute User user) {
+    public ResponseEntity<Message> updateUser(@ModelAttribute User user) {
         logger.debug("Start updateUser() for /updateUser");
-        Message message = userService.updateUser(user);
-
-        return new ResponseEntity<Message>(message, HttpStatus.OK);
+        Validation validation = new Validation();
+        Message message = new Message();
+        if (validation.isValid(user)) {
+            message = userService.updateUser(user);
+            return new ResponseEntity<Message>(message, HttpStatus.OK);
+        } else {
+            message.setMessage("empty/null input/s");
+            message.setStatus(false);
+            return new ResponseEntity<Message>(message, HttpStatus.BAD_REQUEST);
+        }
     }
 
     @RequestMapping(value = "/deleteUser/{id}", method = RequestMethod.DELETE)
-    public ResponseEntity<Message> deleteUser(@Valid @PathVariable("id") int id) {
+    public ResponseEntity<Message> deleteUser(@PathVariable("id") int id) {
         logger.debug("Start deleteUser() for /deleteUser");
-        Message message = userService.deleteUser(id);
+        Validation validation = new Validation();
+        Message message = new Message();
+        if (validation.isValid(id)) {
+            message = userService.deleteUser(id);
+            return new ResponseEntity<Message>(message, HttpStatus.OK);
+        } else {
+            message.setMessage("empty/null id");
+            message.setStatus(false);
+            return new ResponseEntity<Message>(message, HttpStatus.BAD_REQUEST);
+        }
+    }
 
-        return new ResponseEntity<Message>(message, HttpStatus.OK);
+    @RequestMapping(value = "/viewUser/{id}", method = RequestMethod.POST)
+    public ResponseEntity viewUser(@PathVariable("id") int id) {
+        logger.debug("Start viewUser() for /viewUser");
+        Validation validation = new Validation();
+        Message message = new Message();
+        if (validation.isValid(id)) {
+            User user = userService.getUserById(id);
+            if (user.getId() > 0) {
+                return new ResponseEntity<User>(user, HttpStatus.OK);
+            } else {
+                message.setMessage("no user found");
+                message.setStatus(false);
+                return new ResponseEntity<Message>(message, HttpStatus.NOT_FOUND);
+            }
+        } else {
+            message.setMessage("empty/null id");
+            message.setStatus(false);
+            return new ResponseEntity<Message>(message, HttpStatus.BAD_REQUEST);
+        }
     }
 
     @RequestMapping(value = "/listUsers", method = RequestMethod.POST)
@@ -60,11 +103,4 @@ public class UserController {
         return new ResponseEntity<List<User>>(listUsers, HttpStatus.OK);
     }
 
-    @RequestMapping(value = "/viewUser/{id}", method = RequestMethod.POST)
-    public ResponseEntity<User> viewUser(@Valid @PathVariable("id") int id) {
-        logger.debug("Start viewUser() for /viewUser");
-        User user = userService.getUserById(id);
-
-        return new ResponseEntity<User>(user, HttpStatus.OK);
-    }
 }
